+++ Tests/Functional/RenderingTest.php
@@ -1,12 +1,15 @@
 <?php
 namespace T3v\T3vDataMapper\Tests\Functional;
 
-use \TYPO3\CMS\Core\Tests\Functional\Framework\Frontend\Response;
-use \TYPO3\CMS\Core\Tests\FunctionalTestCase;
 use \TYPO3\CMS\Core\Utility\GeneralUtility;
+
+use \Nimut\TestingFramework\Http\Response;
+use \Nimut\TestingFramework\TestCase\FunctionalTestCase;
 
 /**
  * Rendering Test Class
+ *
+ * @package T3v\T3vDataMapper\Tests\Functional
  */
 class RenderingTest extends FunctionalTestCase {
   /**
@@ -72,7 +75,7 @@ class RenderingTest extends FunctionalTestCase {
     if (property_exists($this, 'instancePath')) {
       $instancePath = $this->instancePath;
     } else {
-      $instancePath = ORIGINAL_ROOT . 'typo3temp/functional-' . substr(sha1(get_class($this)), 0, 7);
+      $instancePath = $this->getInstancePath();
     }
 
     $arguments = [
@@ -80,8 +83,12 @@ class RenderingTest extends FunctionalTestCase {
       'requestUrl'   => 'http://localhost' . $requestUrl
     ];
 
-    $template = new \Text_Template(ORIGINAL_ROOT . 'typo3/sysext/core/Tests/Functional/Fixtures/Frontend/request.tpl');
-    $template->setVar(['arguments' => var_export($arguments, true), 'originalRoot' => ORIGINAL_ROOT]);
+    $template = new \Text_Template('ntf://Frontend/Request.tpl');
+    $template->setVar([
+      'arguments'    => var_export($arguments, true),
+      'originalRoot' => ORIGINAL_ROOT,
+      'ntfRoot'      => __DIR__ . '/../../.Build/vendor/nimut/testing-framework/'
+    ]);
 
     $factory = \PHPUnit_Util_PHP::factory();
 
@@ -90,7 +97,7 @@ class RenderingTest extends FunctionalTestCase {
     $result = json_decode($response['stdout'], true);
 
     if ($result === null) {
-      $this->fail('Frontend Response is empty.');
+      $this->fail('Frontend Response is empty:' . LF . 'Error: ' . LF . $response['stderr']);
     }
 
     if ($result['status'] === Response::STATUS_Failure && $failOnFailure) {