+++ Tests/Unit/FirstClassTest.php
@@ -7,6 +7,8 @@ use \T3v\T3vDataMapper\Tests\Unit\Fixtures\LoadableClass;
 
 /**
  * First Class Test Class
+ *
+ * @package T3v\T3vDataMapper\Tests\Unit
  */
 class FirstClassTest extends UnitTestCase {
   /**
@@ -26,6 +28,6 @@ class FirstClassTest extends UnitTestCase {
    * @test
    */
   public function viewHelperBaseClassIsLoadable() {
-    $this->assertTrue(class_exists('TYPO3\\CMS\\Fluid\\Tests\\Unit\\ViewHelpers\\ViewHelperBaseTestcase'));
+    $this->assertTrue(class_exists('TYPO3\CMS\Fluid\Tests\Unit\ViewHelpers\ViewHelperBaseTestcase'));
   }
 }