+++ Tests/Unit/Fixtures/LoadableClass.php
@@ -3,6 +3,8 @@ namespace T3v\T3vDataMapper\Tests\Unit\Fixtures;
 
 /**
  * Loadable Class
+ *
+ * @package T3v\T3vDataMapper\Tests\Unit\Fixtures
  */
 class LoadableClass {
   public function returnsTrue() {